moves bleve index metrics to search package
