+++ pkg/storage/unified/resource/search.go
@@ -9,6 +9,7 @@ import (
 	"time"
 
 	"github.com/grafana/grafana/pkg/cmd/grafana-cli/logger"
+	"github.com/grafana/grafana/pkg/storage/unified/search"
 	"github.com/hashicorp/golang-lru/v2/expirable"
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/trace"
@@ -214,8 +215,8 @@ func (s *searchSupport) init(ctx context.Context) error {
 	}()
 
 	end := time.Now().Unix()
-	if IndexMetrics != nil {
-		IndexMetrics.IndexCreationTime.WithLabelValues().Observe(float64(end - start))
+	if search.IndexMetrics != nil {
+		search.IndexMetrics.IndexCreationTime.WithLabelValues().Observe(float64(end - start))
 	}
 
 	return nil
@@ -260,7 +261,7 @@ func (s *searchSupport) handleEvent(ctx context.Context, evt *WrittenEvent) {
 			return
 		}
 		if evt.Type == WatchEvent_ADDED {
-			IndexMetrics.IndexedKinds.WithLabelValues(evt.Key.Resource).Inc()
+			search.IndexMetrics.IndexedKinds.WithLabelValues(evt.Key.Resource).Inc()
 		}
 	case WatchEvent_DELETED:
 		err = index.Delete(evt.Key)
@@ -268,7 +269,7 @@ func (s *searchSupport) handleEvent(ctx context.Context, evt *WrittenEvent) {
 			s.log.Warn("error deleting document watch event", "error", err)
 			return
 		}
-		IndexMetrics.IndexedKinds.WithLabelValues(evt.Key.Resource).Dec()
+		search.IndexMetrics.IndexedKinds.WithLabelValues(evt.Key.Resource).Dec()
 	default:
 		// do nothing
 		s.log.Warn("unknown watch event", "type", evt.Type)
@@ -279,8 +280,8 @@ func (s *searchSupport) handleEvent(ctx context.Context, evt *WrittenEvent) {
 	if latencySeconds > 5 {
 		logger.Warn("high index latency", "latency", latencySeconds)
 	}
-	if IndexMetrics != nil {
-		IndexMetrics.IndexLatency.WithLabelValues(evt.Key.Resource).Observe(latencySeconds)
+	if search.IndexMetrics != nil {
+		search.IndexMetrics.IndexLatency.WithLabelValues(evt.Key.Resource).Observe(latencySeconds)
 	}
 }
 
@@ -360,8 +361,8 @@ func (s *searchSupport) build(ctx context.Context, nsr NamespacedResource, size 
 	if err != nil {
 		s.log.Warn("error getting doc count", "error", err)
 	}
-	if IndexMetrics != nil {
-		IndexMetrics.IndexedKinds.WithLabelValues(key.Resource).Add(float64(docCount))
+	if search.IndexMetrics != nil {
+		search.IndexMetrics.IndexedKinds.WithLabelValues(key.Resource).Add(float64(docCount))
 	}
 
 	if err != nil {
